snippets/serializers.py
# 外部ライブラリ
from rest_framework import serializers

# 独自ライブラリ
from snippets.models import Snippet


class SnippetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Snippet
        fields = ["id", "title", "code", "linenos", "language", "style"]
